+++ rbldnsd/CHANGELOG.md
@@ -1,6 +1,7 @@
 0.18
 
 * Version bump for new base image
+* Adjust for php-fpm/php_fpm in RC service
 
 ---